+++ docs/source/changelog.rst
@@ -7,6 +7,8 @@ Version 1.0.0 (not yet released)
 * changed: **requires Python 3.5+ now**
 * changed: specifying trackers is now optional with the new ``-t`` switch, thus **trackerless torrents are now
   possible**
+* added: it is now possible to add the best N trackers from `ngosang/trackerslist <https://github.com/ngosang/trackerslist>`_
+  by using the new `bestN shortcut <user.html#automatically-add-the-best-trackers>`__
 * added: DHT bootstrap nodes can now be specified with the new ``--node`` switch (doing so is recommended for
   trackerless torrents)
 * added: **webseed support** with the new ``--webseed`` switch (GetRight style, i.e. `<http://bittorrent.org/beps/bep_0019.html>`_)